refactor: infrastructure 패키지 내 하위 패키지별 정리
