+++ backend/src/main/java/com/woowacourse/kkogkkog/application/MemberService.java
@@ -17,7 +17,7 @@ import com.woowacourse.kkogkkog.domain.repository.MemberRepository;
 import com.woowacourse.kkogkkog.domain.repository.WorkspaceUserRepository;
 import com.woowacourse.kkogkkog.exception.member.MemberHistoryNotFoundException;
 import com.woowacourse.kkogkkog.exception.member.MemberNotFoundException;
-import com.woowacourse.kkogkkog.infrastructure.SlackUserInfo;
+import com.woowacourse.kkogkkog.infrastructure.dto.SlackUserInfo;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
