+++ backend/src/main/java/com/woowacourse/kkogkkog/auth/application/AuthService.java
@@ -5,9 +5,9 @@ import com.woowacourse.kkogkkog.auth.application.dto.MemberCreateResponse;
 import com.woowacourse.kkogkkog.auth.application.dto.TokenResponse;
 import com.woowacourse.kkogkkog.domain.Workspace;
 import com.woowacourse.kkogkkog.domain.repository.WorkspaceRepository;
-import com.woowacourse.kkogkkog.infrastructure.SlackClient;
-import com.woowacourse.kkogkkog.infrastructure.SlackUserInfo;
-import com.woowacourse.kkogkkog.infrastructure.WorkspaceResponse;
+import com.woowacourse.kkogkkog.infrastructure.application.SlackClient;
+import com.woowacourse.kkogkkog.infrastructure.dto.SlackUserInfo;
+import com.woowacourse.kkogkkog.infrastructure.dto.WorkspaceResponse;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
